+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Authoritative changelog in README.md.
 ## Next version
 - Added global configuration.
 - Added UUID generation.
+- Allow changing the name of the field used for the random id.
 
 ## Version 1.0.1 (Dec 22, 2014)
 - Added support for Rails 4.2.